Update README.md
The analysis with METASPLIOT was conducted as part of an investigation to detect and identify vulnerabilities in the target systems network. The tool was used to scan the target system IP address and ports in the network utilizing Nmap with the -sC, -sV, and -vv options.
The project aimed to assess the security  posture of the  network and to detect and identify vulnerabilities in the target system. Nmap through METASPLOIT was employed as a tool to scan the IP address and identify open ports in a network and to detect installed applications